+++ Cargo.toml
@@ -13,10 +13,11 @@ members = [
     "tezos/interop",
     "tezos/encoding",
     "tezos/encoding-derive",
-    "tezos/client",
     "tezos/wrapper",
     "tezos/context",
     "tezos/context-api",
+    "tezos/context-ipc-client",
+    "tezos/protocol-ipc-messages",
     "tezos/spsc",
     "tezos/timing",
     "tezos/lib_storage_ffi",
