+++ protocol_runner/Cargo.toml
@@ -18,7 +18,6 @@ ipc = { path = "../ipc" }
 tezos_api = { path = "../tezos/api" }
 tezos_context = { path = "../tezos/context" }
 tezos_context_api = { path = "../tezos/context-api" }
-tezos_client = { path = "../tezos/client" }
 tezos_interop = { path = "../tezos/interop" }
 tezos_wrapper = { path = "../tezos/wrapper" }
 tezos_messages = { path = "../tezos/messages" }
