+++ protocol_runner/src/main.rs
@@ -6,8 +6,6 @@
 //! problems, from panics to high memory usage, for better stability, we separated protocol into
 //! self-contained process communicating through Unix Socket.
 
-mod ipc_loop;
-
 use clap::{App, Arg};
 use slog::*;
 
@@ -81,7 +79,7 @@ fn main() {
     let shutdown_callback = |log: &Logger| {
         debug!(log, "Shutting down OCaml runtime");
         match std::panic::catch_unwind(|| {
-            tezos_client::client::shutdown_runtime();
+            tezos_interop::shutdown();
         }) {
             Ok(_) => debug!(log, "OCaml runtime shutdown was successful"),
             Err(e) => {
@@ -101,148 +99,16 @@ fn main() {
     }
 
     // Process commands from from the Rust node. Most commands are instructions for the Tezos protocol
-    if let Err(err) = ipc_loop::process_protocol_commands::<crate::tezos::NativeTezosLib, _, _>(
-        cmd_socket_path,
-        &log,
-        shutdown_callback,
-    ) {
-        error!(log, "Error while processing protocol commands"; "reason" => format!("{:?}", err));
-        shutdown_callback(&log);
-    }
+    //if let Err(err) = ipc_loop::process_protocol_commands::<crate::tezos::NativeTezosLib, _, _>(
+    //    cmd_socket_path,
+    //    &log,
+    //    shutdown_callback,
+    //) {
+    //    error!(log, "Error while processing protocol commands"; "reason" => format!("{:?}", err));
+    //    shutdown_callback(&log);
+    //}
+    // TODO: error handling
+    tezos_interop::start_ipc_loop(cmd_socket_path.into());
 
     info!(log, "Protocol runner finished gracefully");
-}
-
-mod tezos {
-    use crypto::hash::{ChainId, ContextHash, ProtocolHash};
-    use tezos_api::ffi::{
-        ApplyBlockError, ApplyBlockRequest, ApplyBlockResponse, BeginApplicationError,
-        BeginApplicationRequest, BeginApplicationResponse, BeginConstructionError,
-        BeginConstructionRequest, CommitGenesisResult, ComputePathError, ComputePathRequest,
-        ComputePathResponse, FfiJsonEncoderError, GetDataError, HelpersPreapplyBlockRequest,
-        HelpersPreapplyError, HelpersPreapplyResponse, InitProtocolContextResult,
-        PrevalidatorWrapper, ProtocolDataError, ProtocolRpcError, ProtocolRpcRequest,
-        ProtocolRpcResponse, RustBytes, TezosRuntimeConfiguration, TezosRuntimeConfigurationError,
-        TezosStorageInitError, ValidateOperationError, ValidateOperationRequest,
-        ValidateOperationResponse,
-    };
-    use tezos_client::client::*;
-    use tezos_context_api::TezosContextConfiguration;
-    use tezos_messages::p2p::encoding::operation::Operation;
-    use tezos_wrapper::protocol::ProtocolApi;
-
-    pub struct NativeTezosLib;
-
-    impl ProtocolApi for NativeTezosLib {
-        fn apply_block(request: ApplyBlockRequest) -> Result<ApplyBlockResponse, ApplyBlockError> {
-            apply_block(request)
-        }
-
-        fn begin_application(
-            request: BeginApplicationRequest,
-        ) -> Result<BeginApplicationResponse, BeginApplicationError> {
-            begin_application(request)
-        }
-
-        fn begin_construction(
-            request: BeginConstructionRequest,
-        ) -> Result<PrevalidatorWrapper, BeginConstructionError> {
-            begin_construction(request)
-        }
-
-        fn validate_operation(
-            request: ValidateOperationRequest,
-        ) -> Result<ValidateOperationResponse, ValidateOperationError> {
-            validate_operation(request)
-        }
-
-        fn call_protocol_rpc(
-            request: ProtocolRpcRequest,
-        ) -> Result<ProtocolRpcResponse, ProtocolRpcError> {
-            call_protocol_rpc(request)
-        }
-
-        fn helpers_preapply_operations(
-            request: ProtocolRpcRequest,
-        ) -> Result<HelpersPreapplyResponse, HelpersPreapplyError> {
-            helpers_preapply_operations(request)
-        }
-
-        fn helpers_preapply_block(
-            request: HelpersPreapplyBlockRequest,
-        ) -> Result<HelpersPreapplyResponse, HelpersPreapplyError> {
-            helpers_preapply_block(request)
-        }
-
-        fn compute_path(
-            request: ComputePathRequest,
-        ) -> Result<ComputePathResponse, ComputePathError> {
-            compute_path(request)
-        }
-
-        fn change_runtime_configuration(
-            settings: TezosRuntimeConfiguration,
-        ) -> Result<(), TezosRuntimeConfigurationError> {
-            change_runtime_configuration(settings)
-        }
-
-        fn init_protocol_context(
-            context_config: TezosContextConfiguration,
-        ) -> Result<InitProtocolContextResult, TezosStorageInitError> {
-            init_protocol_context(context_config)
-        }
-
-        fn genesis_result_data(
-            genesis_context_hash: &ContextHash,
-            chain_id: &ChainId,
-            genesis_protocol_hash: &ProtocolHash,
-            genesis_max_operations_ttl: u16,
-        ) -> Result<CommitGenesisResult, GetDataError> {
-            genesis_result_data(
-                genesis_context_hash,
-                chain_id,
-                genesis_protocol_hash,
-                genesis_max_operations_ttl,
-            )
-        }
-
-        fn assert_encoding_for_protocol_data(
-            protocol_hash: ProtocolHash,
-            protocol_data: Vec<u8>,
-        ) -> Result<(), ProtocolDataError> {
-            assert_encoding_for_protocol_data(protocol_hash, protocol_data)
-        }
-
-        fn apply_block_result_metadata(
-            context_hash: ContextHash,
-            metadata_bytes: RustBytes,
-            max_operations_ttl: i32,
-            protocol_hash: ProtocolHash,
-            next_protocol_hash: ProtocolHash,
-        ) -> Result<String, FfiJsonEncoderError> {
-            apply_block_result_metadata(
-                context_hash,
-                metadata_bytes,
-                max_operations_ttl,
-                protocol_hash,
-                next_protocol_hash,
-            )
-        }
-
-        fn apply_block_operations_metadata(
-            chain_id: ChainId,
-            operations: Vec<Vec<Operation>>,
-            operations_metadata_bytes: Vec<Vec<RustBytes>>,
-            protocol_hash: ProtocolHash,
-            next_protocol_hash: ProtocolHash,
-        ) -> Result<String, FfiJsonEncoderError> {
-            apply_block_operations_metadata(
-                chain_id,
-                operations,
-                operations_metadata_bytes,
-                protocol_hash,
-                next_protocol_hash,
-            )
-        }
-    }
 }
