+++ tezos/client/benches/bench_apply_first_three_blocks.rs
@@ -3,6 +3,8 @@
 #![feature(test)]
 extern crate test;
 
+// TODO: move these benchmarks somewhere else
+
 use std::time::Instant;
 use test::Bencher;
