+++ tezos/context-ipc-client/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "tezos_context_ipc_client"
+version = "1.8.0"
+edition = "2018"
+authors = ["Bruno Deferrari <bruno.deferrari@viablesystems.io>"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0"
+# local dependencies
+crypto = { path = "../../crypto" }
+tezos_context_api = { path = "../context-api" }
+tezos_wrapper = { path = "../wrapper" }
