+++ tezos/interop/src/ipc_message_encoding.rs
@@ -3,27 +3,45 @@
 
 use ocaml_interop::{ocaml_export, OCaml, OCamlBytes, OCamlRef, OCamlRuntime, ToOCaml};
 use tezos_conv::*;
-use tezos_wrapper::service::{NodeMessage, ProtocolMessage};
+use tezos_protocol_ipc_messages::{NodeMessage, ProtocolMessage};
 
-fn encode<'gc>(
+fn encode_response<'gc>(
     cr: &'gc mut OCamlRuntime,
     response: OCamlRef<OCamlNodeMessage>,
-) -> OCaml<'gc, OCamlBytes> {
+) -> OCaml<'gc, Result<OCamlBytes, String>> {
     let rust_response: NodeMessage = response.to_rust(cr);
-    let bytes = bincode::serialize(&rust_response).unwrap(); // TODO: remove unwrap
+    let bytes =
+        bincode::serialize(&rust_response).map_err(|err| format!("Serialization failure: {}", err));
     bytes.to_ocaml(cr)
 }
 
-fn decode<'gc>(
+fn decode_request<'gc>(
     cr: &'gc mut OCamlRuntime,
     bytes: OCamlRef<OCamlBytes>,
 ) -> OCaml<'gc, Result<OCamlProtocolMessage, String>> {
     let ocaml_bytes = cr.get(bytes);
     let rust_bytes = ocaml_bytes.as_bytes();
-    let result: Result<ProtocolMessage, &str> =
-        bincode::deserialize(rust_bytes).map_err(|_| "Deserialization failure");
+    let result = bincode::deserialize::<ProtocolMessage>(rust_bytes)
+        .map_err(|err| format!("Deserialization failure: {}", err));
 
     result.to_ocaml(cr)
+}
+
+#[link(name = "tezos")]
+extern "C" {
+    pub fn initialize_tezedge_ipc_callbacks(
+        tezedge_ipc_encode_response: unsafe extern "C" fn(isize) -> isize,
+        tezedge_ipc_decode_request: unsafe extern "C" fn(isize) -> isize,
+    );
+}
+
+pub fn initialize_callbacks() {
+    unsafe {
+        initialize_tezedge_ipc_callbacks(
+            tezedge_encode_ipc_response_to_ocaml_bytes,
+            tezedge_decode_ipc_message_to_ocaml,
+        );
+    }
 }
 
 // TODO: init function to set callbacks
@@ -32,15 +50,14 @@ ocaml_export! {
     fn tezedge_encode_ipc_response_to_ocaml_bytes(
         cr,
         response: OCamlRef<OCamlNodeMessage>
-    ) -> OCaml<OCamlBytes> {
-        encode(cr, response)
+    ) -> OCaml<Result<OCamlBytes, String>> {
+        encode_response(cr, response)
     }
 
     fn tezedge_decode_ipc_message_to_ocaml(
         cr,
         bytes: OCamlRef<OCamlBytes>
     ) -> OCaml<Result<OCamlProtocolMessage, String>>{
-        decode(cr, bytes)
+        decode_request(cr, bytes)
     }
-
 }
