+++ tezos/protocol-ipc-messages/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "tezos_protocol_ipc_messages"
+version = "1.8.0"
+authors = ["Bruno Deferrari <bruno.deferrari@viablesystems.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+strum_macros = "0.20"
+# local dependencies
+crypto = { path = "../../crypto" }
+tezos_context_api = { path = "../context-api" }
+tezos_api = { path = "../api" }
+tezos_messages = { path = "../messages" }
